requirements.txt
flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.0
python-dotenv==1.0.1
bcrypt==4.1.2
flask-limiter==3.5.0
